+++ cinema_project/src/cinema_project/view/Menu.java
@@ -1,14 +1,9 @@
 package cinema_project.view;
 
-import java.util.Scanner;
-
-import cinema_project.controller.Controller;
-
 public class Menu {
-	 private Scanner sc = new Scanner(System.in);
-	    private Controller userController = new Controller();
 
-	    public void mainMenu() {
+	public static void main(String[] args) {
+		public void mainMenu1() {
 	        while (true) {
 	            System.out.println("=== 10nema ===");
 	            System.out.println("1. 회원가입");
@@ -30,46 +25,6 @@ public class Menu {
 	            }
 	        }
 	    }
-	    private void searchUserInfossnph() {
-	    	System.out.println("=== 아이디,비밀번호 찾기 ===");
-	    	System.out.print("이름 : ");
-	    	String userName = sc.nextLine();
-	    	System.out.print("주민번호 : ");
-	    	String userSsn = sc.nextLine();
-	    	System.out.print("전화번호 : ");
-	    	String userPhone = sc.nextLine();
-	    	
-	    	userController.searchUserInfossnph(userName,userSsn,userPhone);
-	    }
-	    private void createUser() {
-	        System.out.println("=== 회원가입 ===");
-	        System.out.print("아이디: ");
-	        String userId = sc.nextLine();
-	        System.out.print("비밀번호: ");
-	        String userPw = sc.nextLine();
-	        System.out.print("이름: ");
-	        String userName = sc.nextLine();
-	        System.out.print("주민등록번호: ");
-	        String userSsn = sc.nextLine();
-	        System.out.print("핸드폰 번호: ");
-	        String userPhone = sc.nextLine();
-	        
-	        userController.createUser(userId, userPw, userName, userSsn, userPhone);
-	        
-	    }
-	    private void loginUser() {
-	        System.out.println("=== 로그인 ===");
-	        System.out.print("아이디: ");
-	        String userId = sc.nextLine();
-	        System.out.print("비밀번호: ");
-	        String userPw = sc.nextLine();
-
-	        userController.loginUser(userId, userPw);
-	    }
-	  
-	    
-	    public void goodee() {}
-	    
 	}
-	
 
+}
